+++ src/components/App.js
@@ -1,80 +1,18 @@
-import React, { useReducer, useState } from "react"
+import React, { useReducer } from "react"
 import "bootstrap/dist/css/bootstrap.min.css"
+import EventForm from "./EventForm"
 import Event from "./Event"
 import reducer from "../reducers"
 
 const App = () => {
   // 状態遷移したいタイミングでdispatchを呼んで、dispatchの引数にactionを渡す。[]を初期値とする。
+  // App.jsのstateとEventFormのstateは別物になっている
   const [state, dispatch] = useReducer(reducer, [])
-  const [title, setTitle] = useState("")
-  const [body, setBody] = useState("")
-
-  const addEvent = (e) => {
-    // ページをリロードしないようにする
-    e.preventDefault()
-    // dispatchを呼ぶ
-    dispatch({
-      type: "CREATE_EVENT",
-      title,
-      body,
-    })
-
-    setTitle("")
-    setBody("")
-  }
-
-  const deleteAllEvents = (e) => {
-    e.preventDefault()
-    const result = window.confirm(
-      "全てのイベントを本当に削除しても良いですか？"
-    )
-    if (result) dispatch({ type: "DELETE_ALL_EVENTS" })
-  }
-
-  const unCreatable = title === "" || body === ""
+  console.log(state, "in App.js")
 
   return (
     <div className="container-fluid">
-      <h4>イベント作成フォーム</h4>
-      <form>
-        <div className="form-group">
-          <label htmlFor="formEventTitle">タイトル</label>
-          <input
-            type="text"
-            className="form-control"
-            id="formEventTitle"
-            value={title}
-            onChange={(e) => setTitle(e.target.value)}
-          />
-        </div>
-
-        <div className="form-group">
-          <label htmlFor="formEventBody">ボディー</label>
-          <textarea
-            type="text"
-            className="form-control"
-            id="formEventBody"
-            value={body}
-            onChange={(e) => setBody(e.target.value)}
-          />
-        </div>
-
-        <button
-          className="btn btn-primary"
-          onClick={addEvent}
-          disabled={unCreatable}
-        >
-          イベントを作成する
-        </button>
-        <button
-          className="btn btn-danger"
-          onClick={deleteAllEvents}
-          disabled={state.length === 0}
-        >
-          全てのイベントを削除する
-        </button>
-      </form>
-
+      <EventForm />
       <h4>イベント一覧</h4>
       <table className="table table-hover">
         <thead>
